rename akka support + moved method to other directive
