+++ rest/rest-sqs/src/main/scala/org/elasticmq/rest/sqs/GetQueueUrlDirectives.scala
@@ -8,7 +8,7 @@ import spray.json.DefaultJsonProtocol._
 import spray.json.RootJsonFormat
 import org.elasticmq.rest.sqs.model.RequestPayload
 
-trait GetQueueUrlDirectives { this: ElasticMQDirectives with QueueURLModule with AkkaSupport =>
+trait GetQueueUrlDirectives { this: ElasticMQDirectives with QueueURLModule with ResponseMarshaller =>
   def getQueueUrl(p: RequestPayload)(implicit protocol: AWSProtocol, xmlNsVersion: XmlNsVersion) = {
     p.action(GetQueueUrl) {
       rootPath {
