+++ rest/rest-sqs/src/main/scala/org/elasticmq/rest/sqs/ResponseMarshaller.scala
@@ -5,16 +5,11 @@ import spray.json.RootJsonFormat
 import akka.http.scaladsl.marshalling.Marshaller
 import akka.http.scaladsl.model.ContentTypes._
 import akka.http.scaladsl.model.{HttpEntity, RequestEntity}
-import akka.http.scaladsl.server
-import akka.http.scaladsl.server.Directives.complete
 import akka.util.ByteString
-import org.elasticmq.rest.sqs.Constants.EmptyRequestId
 import org.elasticmq.rest.sqs.directives.RespondDirectives
-
-import scala.xml._
 import scala.xml.{Null, UnprefixedAttribute}
 
-trait AkkaSupport {
+trait ResponseMarshaller {
   _: RespondDirectives =>
 
   private def namespace[T](
@@ -39,20 +34,4 @@ trait AkkaSupport {
           }
         }
     }
-
-  def emptyResponse(xmlTagName: String)(implicit protocol: AWSProtocol): server.Route = {
-    protocol match {
-      case AWSProtocol.`AWSJsonProtocol1.0` => complete(200, HttpEntity.Empty)
-      case _ =>
-        respondWith {
-          <wrapper>
-            <ResponseMetadata>
-              <RequestId>
-                {EmptyRequestId}
-              </RequestId>
-            </ResponseMetadata>
-          </wrapper> % Attribute(None, "name", Text(xmlTagName), Null)
-        }
-    }
-  }
 }
