+++ rest/rest-sqs/src/main/scala/org/elasticmq/rest/sqs/directives/RespondDirectives.scala
@@ -5,7 +5,11 @@ import akka.http.scaladsl.model.HttpEntity
 import akka.http.scaladsl.server.{Directives, RequestContext, Route}
 import org.elasticmq.rest.sqs.Constants._
 import scala.language.postfixOps
+import akka.http.scaladsl.server
+import org.elasticmq.rest.sqs.AWSProtocol
+import org.elasticmq.rest.sqs.Constants.EmptyRequestId
 
+import scala.xml._
 import scala.xml.{Elem, Null, UnprefixedAttribute}
 
 trait RespondDirectives {
@@ -36,4 +40,20 @@ trait RespondDirectives {
 
       route(new UnprefixedAttribute("xmlns", "http://queue.amazonaws.com/doc/%s/".format(version), Null))
     }
+
+  def emptyResponse(xmlTagName: String)(implicit protocol: AWSProtocol): server.Route = {
+    protocol match {
+      case AWSProtocol.`AWSJsonProtocol1.0` => complete(200, HttpEntity.Empty)
+      case _ =>
+        respondWith {
+          <wrapper>
+            <ResponseMetadata>
+              <RequestId>
+                {EmptyRequestId}
+              </RequestId>
+            </ResponseMetadata>
+          </wrapper> % Attribute(None, "name", Text(xmlTagName), Null)
+        }
+    }
+  }
 }
